+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.23.19
+### Changed
+- Use set peer timeout instead of constant in block store. 
+
 ## 0.23.18
 ### Changed
 - Allow setting peer timeouts in command line.